Fixed issue with main file

+++ RL_main.py
@@ -134,7 +134,7 @@ episode_rewards = []
 def plot_durations(show_result=False):
     plt.figure(1)
     durations_t = torch.tensor(episode_durations, dtype=torch.float)
-    rewards_t = torch.tensor(episode_rewards, dtype=torch.int32)
+    rewards_t = torch.tensor(episode_rewards, dtype=torch.float)
     if show_result:
         plt.title('Result')
     else:
